Adds Texture Upload and Custom Uniforms
Users can upload custom textures to apply to their models. They can also
add custom Uniform variables that are float, int, bools, vec2, vec3, vec4,
or sampler2d.

After a texture is uploaded it can be accessed by specifying the filename in
the uniforms tab. In the default text for Uniform there is a default texture
specified as `sampler2D my_texture = "textures/beanie.jpg";`

There is a new snippet showing how to apply a texture that is available
in the Uniforms tab.

The uniforms tab now highlights parsing errors.